Refactor get_javascript_result to return either the raw or DOM result
Webkit1 had access to both the dom and javascript.
This means you could get the string "undefined" via eval_js and an empty string via the dom
when asking for the same thing.

In Webkit2 we only have javascript, but the semantic difference of the return values should be preserved.