use click_and_wait with a small timeout in load.t
Using open directly after a click can lead to problems
where the open gets the load events of the load caused
by the preceding click, and not actually loads what it should.
Changing to click_and_wait fixes this but it occasionaly runs
into the timeout, because the load caused by the click is too fast.