Catch page loads earlier so we can wait for them before opening page
Before is_loading is true or a load-changed signal is sent, a
decide-policy signal is emitted to check whether the navigation should
even be accepted. Hook into this to get notification about a pending
navigation event as soon as possible. Otherwise we could even run into a
deadlock when e.g. JS triggers a form submission, but we don't wait for
that and try to open a page instead.