Improve changing selected option
Throw an exception if we couldn't find the option to select instead of
silently continuing. In addition, inline the code to fire the change
event so it will be guaranteed to run after the selection changed.

We must not wait for the selected value to change because changing the
select may trigger a page load and we'd never be able to read the
select element's value. Instead wait for the change signal to be
processed.