Massively reduce latency when waiting for conditions
While we are in an usleep, we cannot process any events. These events however
may be necessary for making progress, i.e. getting to the condition that we're
actually waiting for. By processing events only every 10ms we were actually
starving the queue and protracted the whole process.

Fix by using a Glib timeout instead and use a blocking wait for new events.
This way the process will be woken up exactly when there's something to do
without delay and without unnecessary checking for new events.

Speeds up a large test suite from over an hour to just 10 minutes.